# backend/notification-service-main/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Notification Service API
  description: A robust notification service supporting email, SMS, and push notifications
  version: 1.0.0
  contact:
    name: API Support
    email: hello@mapmytour.in

servers:
  - url: http://localhost:9090/api/v1
    description: Development server

paths:
  /health:
    get:
      summary: Health check endpoint
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /notification/send:
    post:
      summary: Send a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
      responses:
        '200':
          description: Notification queued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /notification/{id}:
    get:
      summary: Get notification status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Notification ID
      responses:
        '200':
          description: Notification found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '404':
          description: Notification not found

components:
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
        errors:
          type: array
          items:
            type: string

    NotificationRequest:
      type: object
      required:
        - type
        - recipient
        - body
      properties:
        type:
          type: string
          enum: [email, sms, push]
        recipient:
          type: string
        subject:
          type: string
        body:
          type: string
        scheduled_at:
          type: integer
          format: int64
        metadata:
          type: object
          additionalProperties:
            type: string